// src/infra/enums/colors.ts
export enum Colors {
    BLACK = 'BLACK',
    SILVER = 'SILVER',
    LIGHT_GRAY = 'LIGHT_GRAY',
    DARK_GRAY = 'DARK_GRAY',
    PASTEL_BLUE = 'PASTEL_BLUE',
    BEIGE = 'BEIGE',
    PASTEL_RED = 'PASTEL_RED',
    DARK_RED = 'DARK_RED ',
    RED = 'RED',
    LIGHT_BLUE = 'LIGHT_BLUE',
    DARK_PASTEL_BLUE = 'DARK_PASTEL_BLUE',
    LAKE_BLUE = 'LAKE_BLUE',
    DARK_BLUE = 'DARK_BLUE',
    LILAC = 'LILAC',
    PINK = 'PINK',
    VIOLET = 'VIOLET',
    YELLOW = 'YELLOW',
    DARK_PURPLE = 'DARK_PURPLE'
}